put common code from pytrack.py and pysense.py in pic_mcu.py
